Basic `enhance` setup with webpack build and custom loader
We now have a pretty slick setup where the app just requires a small core
dependency, chooses itself which widgets it wants to register with the
router and just includes a `<router-fragment>` to use them. The lazy loading
is done under the hood, the app doesn't need to know anything about how that
works under the hood and the router is blissfully unaware at build time of
which widgets might be registered later. All things considered this should
scale really well while keeping the dependencies focused and simple.