Added pagelets using React, Vue and Mithril.
We'll be adding more of these to explore the best and minimal way to interact
for them on the page. It is not clear right now whether pagelets should be able
to declare dependencies on other pagelets. It feels like the maintainer of the
page should have full control over what pagelets should be allowed to be
displayed on the page. Maybe they merely declare dependencies but the loader
won't act upon them unless the page has included the appropriate registration -
in `dev` builds that should cause a warning.