Collapsed API into a single "rye" module
Removed "rye-pagelet" and "rye-pagelet-registry" wrapper modules, a direct call
on the `Rye` global is way easier. We'll see how the usage of the global evolves
in the future. Maybe we import it but build it as external, this is error-prone,
though... Since we're only having one single module right now it made sense to
rename "rye-core" to plain "rye". Another benefit of these changes is that the
API no longer exposes the internal module loader - if we need that to be public
at some point we can release it with the proper API for the use-case.